Fix error from slow master process start.
Fix an intermittent error that occurs when the master process is slow to start.

Occassionally, the master process does not start within a timing threshold
specified in the `${SPARK_HOME}/sbin/spark-daemon.sh` script. Under those
conditions, the output of `${SPARK_HOME}/sbin/start-master.sh` contains
extra text that is not properly parsed by `spark-start` when locating
the master process log file.

Instead of parsing the output of the `start-master.sh` script to locate the
master process log file, this commit will ignore the output of the script,
and list the contents of the spark logging directory to locate the master
process log file.

Additionally, the intermittent error seems to occur when the system is experiencing
slow response time generally. In order to compensate for the slow response time,
the 1 minute timer that waits for the master process to write it's web UI URL and
spark URL to the log file is increased to 10 minutes and 10 minutes.